Further improvements made
