--- modules/Movecraft/src/main/java/net/countercraft/movecraft/repair/RepairUtils.java
@@ -1,4 +1,0 @@
-package net.countercraft.movecraft.repair;
-
-public class RepairUtils {
-}
